Add HTTP controllers for identity provisioning
- Created package structure for HTTP controllers in identity provisioning.
- Implemented ProvisioningJobController to manage provisioning jobs with routes for create, list, get, run, cancel, and delete operations.
- Added ProxySystemController for managing proxy systems with routes for create, list, get, update, activate, deactivate, and delete operations.
- Developed SourceSystemController to handle source systems with similar CRUD operations.
- Introduced TargetSystemController for managing target systems with corresponding routes.
- Implemented TransformationController for managing transformations, including a test route.
- Updated package.d files to include new controllers and ensure proper module imports.
- Modified CommandResult structure to use uint for error codes.
- Adjusted errorResponse and successResponse functions to accept uint for status codes.

+++ service/source/uim/platform/service/application/dto.d
@@ -13,10 +13,10 @@ struct CommandResult {
   /// A message providing additional information about the command result.
   string message;
   /// An optional error code associated with the command result.
-  size_t code;
+  uint code;
   
   ///
-  this(bool success, string id = "", string message = "", size_t code = 0) {
+  this(bool success, string id = "", string message = "", uint code = 0) {
     this.success = success;
     this.id = id;
     this.message = message;
